pyproject.toml
[project]
name = "kpn-genai"
version = "0.1.0"
description = "Kpn GenAI project"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langgraph>=0.6.6,<0.7.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "langchain-ollama>=0.3.7,<0.4.0",
    "langchain-openai>=0.3.32,<0.4.0",
    "langchain-core>=0.3.0,<0.4.0",
    "fastapi>=0.116.2,<0.117.0",
    "uvicorn>=0.35.0,<0.36.0",
    "python-multipart>=0.0.20,<0.0.21",
    "pydantic>=2.0.0,<3.0.0",
    "pymupdf>=1.24.0,<2.0.0",
    "pillow>=10.0.0,<12.0.0",
    "pyyaml>=6.0.0,<7.0.0",
    "python-dateutil>=2.9.0,<3.0.0",
    "mypy (>=1.19.1,<2.0.0)",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.12,<0.13.0",
    "pytest>=8.0.0,<9.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["yaml", "dateutil", "dateutil.parser"]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
